refactor model subscribe
