+++ src/utils/getSubscribeData.js
@@ -1,11 +1,11 @@
 import 'babel-polyfill';
 import * as Videos_new from '../models/videos_new';
-import findSubscribeId from '../models/findSubscribeId';
+import * as Subscribe from '../models/subscribe';
 
 const getSubscribeData = () => {
   return new Promise(resolve => {
     Videos_new.getRandomThreeVideos().then(returnArr => {
-      findSubscribeId().then(senderIDArr => {
+      Subscribe.findSubscribeId().then(senderIDArr => {
         const senderIDArrLength = senderIDArr.length;
         const count = senderIDArr.length;
         const returnObj = {
