+++ src/utils/receivedMessage.js
@@ -1,7 +1,5 @@
 import saveLogData from '../models/saveLogData';
-import saveSubscribeData from '../models/saveSubscribeData';
-import removeSubscribeId from '../models/removeSubscribeId';
-import updateSubscribeData from '../models/updateSubscribeData';
+import * as Subscribe from '../models/subscribe';
 import * as Videos from '../models/videos';
 import FacebookOP from './facebook';
 
@@ -21,7 +19,7 @@ const receivedMessage = async (event) => {
   if (messageText !== undefined) {
     firstStr = messageText.split('')[0];
     messageText = messageText.replace(/\s/g, '');
-    const isUpdate = updateSubscribeData(senderID, true);
+    const isUpdate = Subscribe.updateSubscribeData(senderID, true);
     if (isUpdate) {
       console.log(`${senderID} 更新 isPushable 成功`);
     } else {
@@ -42,13 +40,13 @@ const receivedMessage = async (event) => {
       });
     }
   } else if (messageText === 'GGinin' || messageText === 'GGININ' || messageText === 'gginin' || messageText === 'Gginin') {
-    saveSubscribeData(senderID).then(str => {
+    Subscribe.saveSubscribeData(senderID).then(str => {
       fb.sendTextMessage(senderID, str);
       const str2 = '想看片請輸入 "PPAV" 3:) \n\n其他搜尋功能🔥\n1. 搜尋番號："# + 番號" \n2. 搜尋女優："% + 女優"\n3. 搜尋片名："@ + 關鍵字"\n4. 搜尋標籤："! + 關鍵字"';
       fb.sendTextMessage(senderID, str2);
     });
   } else if (messageText === 'NoGG' || messageText === 'NOGG' || messageText === 'nogg' || messageText === 'noGG' || messageText === 'Nogg') {
-    removeSubscribeId(senderID).then(str => {
+    Subscribe.removeSubscribeId(senderID).then(str => {
       fb.sendTextMessage(senderID, str);
       const str2 = '想看片請輸入 "PPAV" 3:) \n\n其他搜尋功能🔥\n1. 搜尋番號："# + 番號" \n2. 搜尋女優："% + 女優"\n3. 搜尋片名："@ + 關鍵字"\n4. 搜尋標籤："! + 關鍵字"\n\n訂閱每日推播："GGININ"';
       fb.sendTextMessage(senderID, str2);
